Further separate generic and concrete codeintel code.
