+++ enterprise/cmd/frontend/internal/codeintel/httpapi/generic/iface.go
@@ -1,4 +1,4 @@
-package httpapi
+package generic
 
 import (
 	"context"
