+++ enterprise/cmd/frontend/internal/codeintel/httpapi/generic/observability.go
@@ -1,0 +1,41 @@
+package generic
+
+import (
+	"fmt"
+
+	"github.com/sourcegraph/sourcegraph/internal/metrics"
+	"github.com/sourcegraph/sourcegraph/internal/observation"
+)
+
+type Operations struct {
+	handleEnqueue                  *observation.Operation
+	handleEnqueueSinglePayload     *observation.Operation
+	handleEnqueueMultipartSetup    *observation.Operation
+	handleEnqueueMultipartUpload   *observation.Operation
+	handleEnqueueMultipartFinalize *observation.Operation
+}
+
+func NewOperations(observationContext *observation.Context) *Operations {
+	metrics := metrics.NewREDMetrics(
+		observationContext.Registerer,
+		"uploadutil_http",
+		metrics.WithLabels("op"),
+		metrics.WithCountHelp("Total number of method invocations."),
+	)
+
+	op := func(name string) *observation.Operation {
+		return observationContext.Operation(observation.Op{
+			Name:              fmt.Sprintf("uploadutil.http.%s", name),
+			MetricLabelValues: []string{name},
+			Metrics:           metrics,
+		})
+	}
+
+	return &Operations{
+		handleEnqueue:                  op("HandleEnqueue"),
+		handleEnqueueSinglePayload:     op("handleEnqueueSinglePayload"),
+		handleEnqueueMultipartSetup:    op("handleEnqueueMultipartSetup"),
+		handleEnqueueMultipartUpload:   op("handleEnqueueMultipartUpload"),
+		handleEnqueueMultipartFinalize: op("handleEnqueueMultipartFinalize"),
+	}
+}
