+++ enterprise/cmd/frontend/internal/codeintel/httpapi/observability.go
@@ -3,41 +3,34 @@ package httpapi
 import (
 	"fmt"
 
+	generic "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel/httpapi/generic"
 	"github.com/sourcegraph/sourcegraph/internal/metrics"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 )
 
 type Operations struct {
-	authMiddleware                 *observation.Operation
-	handleEnqueue                  *observation.Operation
-	handleEnqueueSinglePayload     *observation.Operation
-	handleEnqueueMultipartSetup    *observation.Operation
-	handleEnqueueMultipartUpload   *observation.Operation
-	handleEnqueueMultipartFinalize *observation.Operation
+	*generic.Operations
+	authMiddleware *observation.Operation
 }
 
 func NewOperations(observationContext *observation.Context) *Operations {
 	metrics := metrics.NewREDMetrics(
 		observationContext.Registerer,
-		"codeintel_httpapi",
+		"codeintel_httpapi_auth",
 		metrics.WithLabels("op"),
 		metrics.WithCountHelp("Total number of method invocations."),
 	)
 
 	op := func(name string) *observation.Operation {
 		return observationContext.Operation(observation.Op{
-			Name:              fmt.Sprintf("codeintel.httpapi.%s", name),
+			Name:              fmt.Sprintf("codeintel.httpapi.auth.%s", name),
 			MetricLabelValues: []string{name},
 			Metrics:           metrics,
 		})
 	}
 
 	return &Operations{
-		authMiddleware:                 op("authMiddleware"),
-		handleEnqueue:                  op("HandleEnqueue"),
-		handleEnqueueSinglePayload:     op("handleEnqueueSinglePayload"),
-		handleEnqueueMultipartSetup:    op("handleEnqueueMultipartSetup"),
-		handleEnqueueMultipartUpload:   op("handleEnqueueMultipartUpload"),
-		handleEnqueueMultipartFinalize: op("handleEnqueueMultipartFinalize"),
+		Operations:     generic.NewOperations(observationContext),
+		authMiddleware: op("authMiddleware"),
 	}
 }
