+++ enterprise/cmd/frontend/internal/codeintel/httpapi/upload_handler.go
@@ -1,156 +1,113 @@
 package httpapi
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"fmt"
-	"io"
 	"net/http"
 
-	"github.com/opentracing/opentracing-go/log"
-
+	"github.com/sourcegraph/log"
 	sglog "github.com/sourcegraph/log"
 
-	"github.com/sourcegraph/sourcegraph/internal/observation"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel/httpapi/auth"
+	generic "github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel/httpapi/generic"
+	"github.com/sourcegraph/sourcegraph/internal/actor"
+	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/errcode"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
+	"github.com/sourcegraph/sourcegraph/internal/lazyregexp"
 	"github.com/sourcegraph/sourcegraph/internal/uploadstore"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
-type UploadHandler[T any] struct {
-	logger              sglog.Logger
-	dbStore             DBStore[T]
-	uploadStore         uploadstore.Store
-	operations          *Operations
-	metadataFromRequest func(ctx context.Context, r *http.Request) (T, int, error)
-}
+var revhashPattern = lazyregexp.New(`^[a-z0-9]{40}$`)
 
-func NewUploadHandler[T any](
-	logger sglog.Logger,
-	dbStore DBStore[T],
+func NewUploadHandler(
+	db database.DB,
+	dbStore generic.DBStore[UploadMetadata],
 	uploadStore uploadstore.Store,
+	internal bool,
+	authValidators auth.AuthValidatorMap,
 	operations *Operations,
-	metadataFromRequest func(ctx context.Context, r *http.Request) (T, int, error),
 ) http.Handler {
-	handler := &UploadHandler[T]{
-		logger:              logger,
-		dbStore:             dbStore,
-		uploadStore:         uploadStore,
-		operations:          operations,
-		metadataFromRequest: metadataFromRequest,
-	}
+	logger := sglog.Scoped("UploadHandler", "").With(
+		sglog.Bool("internal", internal),
+	)
 
-	return http.HandlerFunc(handler.handleEnqueue)
-}
+	metadataFromRequest := func(ctx context.Context, r *http.Request) (UploadMetadata, int, error) {
+		commit := getQuery(r, "commit")
+		if !revhashPattern.Match([]byte(commit)) {
+			return UploadMetadata{}, http.StatusBadRequest, errors.Errorf("commit must be a 40-character revhash")
+		}
 
-var errUnprocessableRequest = errors.New("unprocessable request: missing expected query arguments (uploadId, index, or done)")
-
-// POST /upload
-//
-// handleEnqueue dispatches to the correct handler function based on the request's query args. Running
-// commands such as `src code-intel upload` will cause one of two sequences of requests to occur. For
-// uploads that are small enough repos (that can be uploaded in one-shot), only one request will be made:
-//
-//   - POST `/upload?{metadata}`
-//
-// where `{metadata}` contains the keys `repositoryId`, `commit`, `root`, `indexerName`, `indexerVersion`,
-// and `associatedIndexId`.
-//
-// For larger uploads, the requests are broken up into a setup request, a serires of upload requests,
-// and a finalization request:
-//
-//   - POST `/upload?multiPart=true,numParts={n},{metadata}`
-//   - POST `/upload?uploadId={id},index={i}`
-//   - POST `/upload?uploadId={id},done=true`
-//
-// See the functions the following functions for details on how each request is handled:
-//
-//   - handleEnqueueSinglePayload
-//   - handleEnqueueMultipartSetup
-//   - handleEnqueueMultipartUpload
-//   - handleEnqueueMultipartFinalize
-func (h *UploadHandler[T]) handleEnqueue(w http.ResponseWriter, r *http.Request) {
-	// Wrap the interesting bits of this in a function literal that's immediately
-	// executed so that we can instrument the duration and the resulting error more
-	// easily. The remainder of the function simply serializes the result to the
-	// HTTP response writer.
-	payload, statusCode, err := func() (_ any, statusCode int, err error) {
-		ctx, trace, endObservation := h.operations.handleEnqueue.With(r.Context(), &err, observation.Args{})
-		defer func() {
-			endObservation(1, observation.Args{LogFields: []log.Field{
-				log.Int("statusCode", statusCode),
-			}})
-		}()
-
-		uploadState, statusCode, err := h.constructUploadState(ctx, r)
+		// Ensure that the repository and commit given in the request are resolvable.
+		repositoryName := getQuery(r, "repository")
+		repositoryID, statusCode, err := ensureRepoAndCommitExist(ctx, logger, db, repositoryName, commit)
 		if err != nil {
-			return nil, statusCode, err
-		}
-		trace.Log(
-			log.Int("uploadID", uploadState.uploadID),
-			log.Int("numParts", uploadState.numParts),
-			log.Int("numUploadedParts", len(uploadState.uploadedParts)),
-			log.Bool("multipart", uploadState.multipart),
-			log.Bool("suppliedIndex", uploadState.suppliedIndex),
-			log.Int("index", uploadState.index),
-			log.Bool("done", uploadState.done),
-			log.Object("metadata", uploadState.metadata),
-		)
-
-		if uploadHandlerFunc := h.selectUploadHandlerFunc(uploadState); uploadHandlerFunc != nil {
-			return uploadHandlerFunc(ctx, uploadState, r.Body)
+			return UploadMetadata{}, statusCode, err
 		}
 
-		return nil, http.StatusBadRequest, errUnprocessableRequest
-	}()
-	if err != nil {
-		if statusCode >= 500 {
-			h.logger.Error("codeintel.httpapi: failed to enqueue payload", sglog.Error(err))
-		}
-
-		http.Error(w, fmt.Sprintf("failed to enqueue payload: %s", err.Error()), statusCode)
-		return
+		// Populate state from request
+		return UploadMetadata{
+			RepositoryID:      repositoryID,
+			Commit:            commit,
+			Root:              sanitizeRoot(getQuery(r, "root")),
+			Indexer:           getQuery(r, "indexerName"),
+			IndexerVersion:    getQuery(r, "indexerVersion"),
+			AssociatedIndexID: getQueryInt(r, "associatedIndexId"),
+		}, 0, nil
 	}
 
-	if payload == nil {
-		// 204 with no body
-		w.WriteHeader(http.StatusNoContent)
-		return
+	handler := generic.NewUploadHandler(
+		logger,
+		dbStore,
+		uploadStore,
+		operations.Operations,
+		metadataFromRequest,
+	)
+
+	if !internal {
+		// 🚨 SECURITY: Non-internal installations of this handler will require a user/repo
+		// visibility check with the remote code host (if enabled via site configuration).
+		handler = auth.AuthMiddleware(handler, db, authValidators, operations.authMiddleware)
 	}
 
-	data, err := json.Marshal(payload)
-	if err != nil {
-		h.logger.Error("codeintel.httpapi: failed to serialize result", sglog.Error(err))
-		http.Error(w, fmt.Sprintf("failed to serialize result: %s", err.Error()), http.StatusInternalServerError)
-		return
-	}
-
-	// 202 with identifier payload
-	w.WriteHeader(http.StatusAccepted)
-
-	if _, err := io.Copy(w, bytes.NewReader(data)); err != nil {
-		h.logger.Error("codeintel.httpapi: failed to write payload to client", sglog.Error(err))
-	}
+	return handler
 }
 
-type uploadHandlerFunc[T any] func(context.Context, uploadState[T], io.Reader) (any, int, error)
+func ensureRepoAndCommitExist(ctx context.Context, logger log.Logger, db database.DB, repoName, commit string) (int, int, error) {
+	// 🚨 SECURITY: Bypass authz here; we've already determined that the current request is
+	// authorized to view the target repository; they are either a site admin or the code
+	// host has explicit listed them with some level of access (depending on the code host).
+	ctx = actor.WithInternalActor(ctx)
 
-func (h *UploadHandler[T]) selectUploadHandlerFunc(uploadState uploadState[T]) uploadHandlerFunc[T] {
-	if uploadState.uploadID == 0 {
-		if uploadState.multipart {
-			return h.handleEnqueueMultipartSetup
+	//
+	// 1. Resolve repository
+
+	repo, err := backend.NewRepos(logger, db).GetByName(ctx, api.RepoName(repoName))
+	if err != nil {
+		if errcode.IsNotFound(err) {
+			return 0, http.StatusNotFound, errors.Errorf("unknown repository %q", repoName)
 		}
 
-		return h.handleEnqueueSinglePayload
+		return 0, http.StatusInternalServerError, err
 	}
 
-	if uploadState.suppliedIndex {
-		return h.handleEnqueueMultipartUpload
+	//
+	// 2. Resolve commit
+
+	if _, err := backend.NewRepos(logger, db).ResolveRev(ctx, repo, commit); err != nil {
+		var reason string
+		if errors.HasType(err, &gitdomain.RevisionNotFoundError{}) {
+			reason = "commit not found"
+		} else if gitdomain.IsCloneInProgress(err) {
+			reason = "repository still cloning"
+		} else {
+			return 0, http.StatusInternalServerError, err
+		}
+
+		logger.Warn("Accepting LSIF upload with unresolvable commit", log.String("reason", reason))
 	}
 
-	if uploadState.done {
-		return h.handleEnqueueMultipartFinalize
-	}
-
-	return nil
+	return int(repo.ID), 0, nil
 }
