+++ mockgen.test.yaml
@@ -34,7 +34,11 @@
   interfaces:
     - githubClient
 - filename: enterprise/cmd/frontend/internal/codeintel/httpapi/mocks_test.go
-  path: github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel/httpapi
+  path: github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel/httpapi/generic
+  interfaces:
+    - DBStore
+- filename: enterprise/cmd/frontend/internal/codeintel/httpapi/generic/mocks_test.go
+  path: github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/codeintel/httpapi/generic
   interfaces:
     - DBStore
 - filename: enterprise/cmd/frontend/internal/codeintel/httpapi/auth/mocks_test.go
